Use Abstract instead of Base as a prefix.
